Vertical scroll to the nearest parent element (#6)
Change that allows to use the Message-list component within a page with other components. Basically, when the page scroll was activated then the component was at the top of the page (not a problem in the sample app provided because the chat is in full window). With this modification, this will no longer be the case, so the scroll will just be in the scrollable div and the component will stay in its place.